auto uploader/xool-main/requirements.txt
requests
Pillow
opennsfw2
requests_toolbelt
imagehash